add new buttons in notification

+++ src/app/components/notification/notification.component.ts
@@ -1,9 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { BehaviorSubject } from 'rxjs';
 import { INotification } from 'src/app/interfaces/notification.interface';
 import { LoginService } from 'src/app/services/login/login.service';
 import { NotificationService } from 'src/app/services/notification/notification.service';
+import { ViewModeService } from 'src/app/services/view-mode/view-mode.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-notification',
@@ -12,6 +13,7 @@ import { NotificationService } from 'src/app/services/notification/notification.
 })
 export class NotificationComponent implements OnInit {
   newNotificationsCount: number = 0;
+  notificationsCount: number = 0;
 
   newNotifications$ = this.notificationService.newNotifications$;
   oldNotifications$ = this.notificationService.oldNotifications$;
@@ -19,44 +21,86 @@ export class NotificationComponent implements OnInit {
   constructor(
     private notificationService: NotificationService,
     private loginService: LoginService,
+    private modeView: ViewModeService,
     private router: Router
   ) {}
 
   ngOnInit(): void {
     const notifications =
       this.notificationService.getNotificationsFromLocalStorage();
+
     this.notificationService.separateNotifications(notifications);
+
     this.notificationService.clearOldNotifications();
+
     this.notificationService.newNotifications$.subscribe((notifications) => {
       this.newNotificationsCount = notifications.length;
     });
+
+    this.notificationService.notifications$.subscribe(
+      (notifications) => (this.notificationsCount = notifications.length)
+    );
   }
 
   /**
    *
    * @param n
    */
-  markRead(n: INotification) {
+  markRead(n: INotification): void {
     this.notificationService.markNotificationAsRead(n);
   }
 
-  deleteNotification(notification: INotification) {
-    this.notificationService.deleteNotification(notification);
+  markAllAsRead(): void {
+    this.notificationService.markAllNotificationAsRead();
+  }
+
+  deleteNotification(notification: INotification): void {
+    Swal.fire({
+      title: 'Eliminar notificación',
+      text: '¿Estas seguro de que quieres eliminar la siguiente notificación?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Eliminar',
+      cancelButtonText: 'Cancelar',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.notificationService.deleteNotification(notification);
+      }
+    });
+  }
+
+  isMarkAllAsRead(): boolean {
+    return this.newNotificationsCount === 0;
+  }
+
+  deleteAllNotification(): void {
+    Swal.fire({
+      title: 'Eliminar notificaciones',
+      text: '¿Estas seguro de que quieres eliminar todas las notificaciones?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Eliminar',
+      cancelButtonText: 'Cancelar',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.notificationService.deleteAllNotification();
+      }
+    });
   }
 
   goToPage(notification: INotification): void {
     const role = this.loginService.getCurrentUserRole();
     this.notificationService.markNotificationAsRead(notification);
+
     notification.page =
       role === 'ADMIN'
         ? '/admin' + notification.page
         : '/user-dashboard' + notification.page;
-    debugger;
+    this.modeView.modeEdit('no');
     this.router.navigate([notification.page]); // Usa el método navigate para redirigir a la página
-  }
-
-  shouldShowButton(page: string): boolean {
-    debugger;
-    return page.trim() !== '';
   }
 }
